Add conceptification section; update references; clean up
Add section on results of investigating constraining template
parameters with concepts, in the manner of P1813R0 with the numeric
algorithms.  We concluded that we disagree with the approach of
P1813R0, and that the Standard's current **GENERALIZED_SUM** approach
better expresses numeric algorithms' behavior.

Update references to the current revision of P0009 (`mdspan`).