Clean up constraints on add & copy
Remove the explicitly stated constraint on `add` and `copy`
that the rank of the array arguments be no more than 2.
This is redundant, because we already impose this
via the existing constraints on template parameters named
`in_object*_t`, `inout_object*_t`, or `out_object*_t`.
If we later wish to relax this restriction,
then we only have to do so in one place.