P1673: Explain why back-end not specified
During LEWG review of P1673R8 on 2022/05/24,
LEWG asked us to add a section to the paper
explaining why we don't define an interface
for customization of the "back-end" optimized BLAS operations.
This section already existed, but we rewrote it to improve clarity.
Please see the section titled
"We do not require using the BLAS library or any particular 'back-end'."

Fixed P1674 links to point to R2.

Fixes #252.