P2642R1 revisions
Change padding stride to function as an overalignment factor
if less than the extent to pad.

Remove mapping constructor that takes extents_type and
extents<index_type, padding_stride>, because the latter
may not be the actual padding stride.

Update the non-wording sections to match the above changes.

Make other minor content and formatting fixes.