P1673R12: triangular_matrix_{left,right}_product
Change `triangular_matrix_{left,right}_product` to
`triangular_matrix_product`,
distinguishing the left and right cases
by order of the `A, t` and `B` (or `C`,
in the case of in-place `triangular_matrix_product`)
parameters.

NOTE: As a result, `triangular_matrix_product`
(for the in-place right product case)
is now an exception to the rule that
output or in/out parameters appear last.