P2642: Fix layout mapping requirements
Change layout mapping requirements so that a layout mapping's
`extents()` member function can return either `extents_type`
or `const extents_type&`, instead of just `const extents_type&`.
This lets `extents()` return a temporary.
The padded layout mappings in this proposal need to do this
so that they only need to store the "inner" extents,
not their actual extents.
(Lifetime extension doesn't apply to a temporary
bound to a return value of a function in a return statement,
so a return type of `const extents_type&`
would always result in a dangling reference.
Storing the actual `extents(``) return value
instead of the inner extents would either waste storage
(2x extents) or complicate the wording.

This commit does NOT rerender the document!