P2642: Fix converting ctors & op==
Change converting constructors from
`layout_{left,right}_padded<ps>::mapping<OE>`,
and `operator==` taking that type,
to use a constraint instead of overloading.
This helps with type deduction in practice.
As a result, we needed a way to get the `padding_stride`
template argument of a mapping.  We found it easiest to do this
by adding a `static constexpr size_t padding_stride`
public member to the mapping.  Another approach would have been
to introduce an exposition-only trait, but we think
the public member would be more generally useful.