P1673: Fix [linalg.algs.reqs] 1
Fix [linalg.algs.reqs] 1 by changing "type requirements" to "Constraints."
Add the Constraint that ExecutionPolicy is an execution policy.
Remove the requirement that the algorithms that take `ExecutionPolicy`
are parallel algorithms, because that would be circular with
**[algorithms.parallel]** 2 ("A _parallel algorithm_ is
a function template listed in this document
with a template parameter named `ExecutionPolicy`").