P1673: Triangle match: Constraint -> Mandate
For `triangular_*`, `symmetric_*`, and `hermitian_*` functions
that take a `Triangle` parameter and
an `mdspan` with `layout_blas_packed` layout,
change the requirement that the layout's `Triangle`
match the function's `Triangle` parameter,
from a Constraint to a Mandate.
This should not result in ambiguous overloads,
since `Triangle` is already Constrained to be
`upper_triangle_t` or `lower_triangle_t`.