P2642: Address LWG 2024/01/17 review feedback
For `layout_left::mapping`'s converting constructor from
`LayoutLeftPaddedMapping`, and for `layout_right::mapping`'s
converting constructor from `LayoutRightPadded`,
simplify the wording per LWG's instructions.

For example, for `layout_left::mapping`'s converting constructor,
LWG asked us to change the "then" clause from the following.

> then `Extents::static_extent(0)` is a multiple of
> `LayoutLeftPaddedMapping::padding_value`

If `LayoutLeftPaddedMapping::`_`static-padding-stride`_
is not `dynamic_extent`, then the input mapping
must have static extent 0 _and_ static `padding_value`.

It's possible for
`LayoutLeftPaddedMapping::extents_type::static_extent(0)`
to be not equal to `dynamic_extent`,
but for `padding_value` to equal `dynamic_extent` --
that is, a compile-time input extent,
but a run-time padding value.
This is what makes the Mandates clause not just
equivalent to the condition that the result extents
are constructible from the input extents.